pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdfkit-cli"
version = "0.1.0"
description = "全能 PDF 命令行处理工具"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["pdf", "cli", "pdf-processing", "ocr"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
]

dependencies = [
    # PDF 处理核心
    "PyMuPDF>=1.23.0",
    "pypdf>=3.0.0",
    "pdfplumber>=0.10.0",
    "pdf2image>=1.16.0",
    "img2pdf>=0.5.0",
    "pikepdf>=8.0.0",

    # OCR (阿里百炼 Qwen3-VL)
    "openai>=1.0.0",
    "Pillow>=10.0.0",

    # 转换
    "python-docx>=1.0.0",
    "weasyprint>=60.0",  # HTML 转 PDF（替代 pdfkit 避免命名冲突）
    "playwright>=1.40.0",

    # CLI 框架
    "typer>=0.9.0",
    "rich>=13.0.0",

    # MCP 服务器
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.25.0",

    # 其他
    "tqdm>=4.65.0",
    "watchdog>=3.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
pdfkit = "pdfkit.__main__:main"
pdfkit-cli = "pdfkit.__main__:main"
pdfkit-mcp = "pdfkit.mcp.server:main"

[project.urls]
Homepage = "https://github.com/your/pdfkit"
Documentation = "https://github.com/your/pdfkit#readme"
Repository = "https://github.com/your/pdfkit"
Issues = "https://github.com/your/pdfkit/issues"

[tool.hatch.build.targets.wheel]
packages = ["pdfkit"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
